@@ -23,7 +23,7 @@ import java.util.Map;
         TranslateItemConfiguration.class})
 @PropertySource(value = "classpath:META-INF/coremedia/gcc-workflow.properties")
 @DefaultAnnotation(NonNull.class)
-public class TranslateGccConfiguration {
+public class TranslateGccAutoConfiguration {
 
   /**
    * A strategy for extracting derived contents from the default translation.xml workflow definition.
